+++ packages/shared/src/album.ts
@@ -4,7 +4,7 @@ export interface Album {
   artistId?: string;
   artistName?: string;
   artists?: string[];
-  albumType?: string;
+  releaseType?: string;
   year?: number;
   genre?: string;
   genreId?: string;